% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.phantom.R
\name{install.phantom}
\alias{install.phantom}
\title{Saves Absolute Path to PhantomJS}
\usage{
install.phantom(path)
}
\arguments{
\item{path}{Path to executable 'phantomJS' binary, downloadble from \url{http://phantomjs.org/download.html}}
}
\value{
The function does not return anything
}
\description{
The \code{install.phantom()} function takes the path to executable
\code{'phantomJS'} binary and memorizes it for future calls. PhantomJS binary
can be downloaded from \url{http://phantomjs.org/download.html}
}
\examples{
\dontrun{
#save the absolute path to phantomJS by giving relative path
install.phantom("./bin/phantomjs")

#save the absolute path to phantomJS
install.phantom("c:\\phantomjs\\bin\\phantomjs")

#convert JPEG to PDF
convertGraph("./example.jpeg", "./example.pdf", path = "path to executable phantomJS" )
}

}
\author{
E. F. Haghish \cr
Medical Informatics and Biostatistics (IMBI) \cr
University of Freiburg, Germany \cr
\email{haghish@imbi.uni-freiburg.de} \cr
\cr
Department of Mathematics and Computer Science \cr
University of Southern Denmark \cr
\email{haghish@imada.sdu.dk}
}
\keyword{PhantomJS}

